pyproject.toml
[project]
name = "deepagents-researcher"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "deepagents>=0.5.3",
    "tavily-python>=0.7.23",
]
